Enable snmpd so it works on systemd systems
With systemd systems it appears that ensure => running is insufficient
to make sure the service is running. You also have to enable the service
so that its unit is startable. Since we pretty much always want snmpd
running just enable it.

This behavior was noticed when git.openstack.org was rebuilt on CentOS 7
which uses systemd.

Change-Id: I44f7f73abd5ecf12dad2522a7c48df50b12f7e9b